Use `GrailsASTUtils.isDomainClass()` instead of `DomainClassArtefactHandler.isDomainClass()`
