+++ grace-plugin-controllers/src/main/groovy/grails/artefact/controller/support/ResponseRedirector.groovy
@@ -34,8 +34,8 @@ import grails.web.mapping.mvc.RedirectEventListener
 import grails.web.mapping.mvc.exceptions.CannotRedirectException
 import grails.web.mvc.FlashScope
 
+import org.grails.compiler.injection.GrailsASTUtils
 import org.grails.core.artefact.ControllerArtefactHandler
-import org.grails.core.artefact.DomainClassArtefactHandler
 import org.grails.datastore.mapping.model.config.GormProperties
 
 /**
@@ -89,7 +89,7 @@ trait ResponseRedirector implements WebAttributes {
     void redirect(object) {
         if (object) {
             Class<?> objectClass = object.getClass()
-            boolean isDomain = (object instanceof GroovyObject) && DomainClassArtefactHandler.isDomainClass(objectClass)
+            boolean isDomain = (object instanceof GroovyObject) && GrailsASTUtils.isDomainClass(objectClass)
             if (isDomain) {
                 def id = ((GroovyObject) object).getProperty(GormProperties.IDENTITY)
                 if (id != null) {
