+++ grace-web-databinding/src/main/groovy/grails/web/databinding/GrailsWebDataBinder.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2023 the original author or authors.
+ * Copyright 2014-2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -45,7 +45,7 @@ import grails.util.GrailsNameUtils
 import grails.validation.DeferredBindingActions
 import grails.validation.ValidationErrors
 
-import org.grails.core.artefact.DomainClassArtefactHandler
+import org.grails.compiler.injection.GrailsASTUtils
 import org.grails.core.exceptions.GrailsConfigurationException
 import org.grails.databinding.IndexedPropertyReferenceDescriptor
 import org.grails.databinding.xml.GPathResultMap
@@ -241,7 +241,7 @@ class GrailsWebDataBinder extends SimpleDataBinder {
     }
 
     protected boolean isDomainClass(final Class<?> clazz) {
-        DomainClassArtefactHandler.isDomainClass(clazz)
+        GrailsASTUtils.isDomainClass(clazz)
     }
 
     protected Object getIdentifierValueFrom(source) {
